requirements.txt
requests>=2.26.0
pyyaml>=6.0
requests_mock>=1.9.3  # For testing
